+++ rust/main/chains/hyperlane-cosmos-native/src/providers.rs
@@ -1,7 +1,0 @@
-mod cosmos;
-mod grpc;
-mod rpc;
-
-pub use cosmos::CosmosNativeProvider;
-pub use grpc::*;
-pub use rpc::*;