+++ typescript/cosmos-sdk/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @hyperlane-xyz/cosmos-sdk
 
+## 18.2.0
+
+### Minor Changes
+
+- dfa9d368c: Added a `getCometClientOrFail` to the `HyperlaneModuleClient` to expose the internal provider connection
+
+### Patch Changes
+
+- @hyperlane-xyz/cosmos-types@18.2.0
+
+## 18.1.0
+
+### Patch Changes
+
+- @hyperlane-xyz/cosmos-types@18.1.0
+
+## 18.0.0
+
+### Patch Changes
+
+- @hyperlane-xyz/cosmos-types@18.0.0
+
 ## 17.0.0
 
 ### Patch Changes